requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
playwright>=1.40.0
croniter>=2.0.5
loguru>=0.7.2
